@@ -1,76 +1,46 @@
 /** @type {import('tailwindcss').Config} */
-module.exports = {
-  darkMode: ["class"],
+export default {
   content: [
-    './pages/**/*.{ts,tsx}',
-    './components/**/*.{ts,tsx}',
-    './app/**/*.{ts,tsx}',
-    './src/**/*.{ts,tsx}',
-	],
+    "./index.html",
+    "./src/**/*.{js,ts,jsx,tsx}",
+  ],
   theme: {
-    container: {
-      center: true,
-      padding: "2rem",
-      screens: {
-        "2xl": "1400px",
-      },
-    },
     extend: {
-      colors: {
-        border: "hsl(var(--border))",
-        input: "hsl(var(--input))",
-        ring: "hsl(var(--ring))",
-        background: "hsl(var(--background))",
-        foreground: "hsl(var(--foreground))",
-        primary: {
-          DEFAULT: "hsl(var(--primary))",
-          foreground: "hsl(var(--primary-foreground))",
-        },
-        secondary: {
-          DEFAULT: "hsl(var(--secondary))",
-          foreground: "hsl(var(--secondary-foreground))",
-        },
-        destructive: {
-          DEFAULT: "hsl(var(--destructive))",
-          foreground: "hsl(var(--destructive-foreground))",
-        },
-        muted: {
-          DEFAULT: "hsl(var(--muted))",
-          foreground: "hsl(var(--muted-foreground))",
-        },
-        accent: {
-          DEFAULT: "hsl(var(--accent))",
-          foreground: "hsl(var(--accent-foreground))",
-        },
-        popover: {
-          DEFAULT: "hsl(var(--popover))",
-          foreground: "hsl(var(--popover-foreground))",
-        },
-        card: {
-          DEFAULT: "hsl(var(--card))",
-          foreground: "hsl(var(--card-foreground))",
-        },
-      },
-      borderRadius: {
-        lg: "var(--radius)",
-        md: "calc(var(--radius) - 2px)",
-        sm: "calc(var(--radius) - 4px)",
-      },
-      keyframes: {
-        "accordion-down": {
-          from: { height: 0 },
-          to: { height: "var(--radix-accordion-content-height)" },
-        },
-        "accordion-up": {
-          from: { height: "var(--radix-accordion-content-height)" },
-          to: { height: 0 },
-        },
-      },
-      animation: {
-        "accordion-down": "accordion-down 0.2s ease-out",
-        "accordion-up": "accordion-up 0.2s ease-out",
-      },
+      // Discrete Rainbow Scheme (23 colors)
+      // Designed by Paul Tol
+      // https://personal.sron.nl/~pault/
+      "light_purple_1": "#e8ecfb",
+      "light_purple_2": "#d9cce3",
+      "purple_4": "#caaccb",
+      "purple_5": "#ba8db4",
+      "purple_7": "#aa6f9e",
+      "dark_purple_8": "#994f88",
+      "dark_purple_9": "#882e72",
+      "dark_blue_10": "#1956b0",
+      "blue_11": "#437dbf",
+      "blue_13": "#6195cf",
+      "light_blue_14": "#7bafde",
+      "dark_green_15": "#4eb265",
+      "green_16": "#90c987",
+      "light_green_17": "#cae0ab",
+      "yellow_18": "#f7f056",
+      "light_orange_19": "#f7cb45",
+      "orange_23": "#ee8026",
+      "dark_orange_25": "#e65518",
+      "red_26": "#dc050c",
+      "red_27": "#a5170e",
+      "dark_red_28": "#72190e",
+      "dark_red_29": "#42150a",
+      "gray": "#777777",
     },
   },
-  plugins: [require("tailwindcss-animate")],
+  plugins: [],
 }
+
+
+
+
+
+
+
+
