@@ -1,9 +1,10 @@
-import { Button } from "@/components/ui/button";
+function App() {
 
-export default function App() {
   return (
-    <div>
-      <Button>Click me</Button>
+    <div className="bg-gray-700 min-h-screen w-full">
+
     </div>
-  );
+  )
 }
+
+export default App
