@@ -1,12 +1,7 @@
-import path from "path"
-import react from '@vitejs/plugin-react-swc'
-import { defineConfig } from "vite"
+import { defineConfig } from 'vite'
+import react from '@vitejs/plugin-react'
 
+// https://vitejs.dev/config/
 export default defineConfig({
   plugins: [react()],
-  resolve: {
-    alias: {
-      "@": path.resolve(__dirname, "./@"),
-    },
-  },
 })
